+++ tests/unit/validators/ds-error-test.js
@@ -1,10 +1,22 @@
 import EmberObject from '@ember/object';
-import DS from 'ember-data';
 import { module, test } from 'qunit';
 import validate from 'ember-validators/ds-error';
 import processResult from '../../helpers/process-result';
 
 let model, result;
+
+class DSErrors extends Map {
+  errorsFor(key) {
+    let errors = [];
+    for (let [k, val] of this.entries()) {
+      if (k === key) {
+        errors.push({ message: val });
+      }
+    }
+
+    return errors;
+  }
+}
 
 module('Unit | Validator | ds-error');
 
@@ -21,14 +33,14 @@ test('it works', function (assert) {
   assert.expect(2);
 
   model = EmberObject.create({
-    errors: DS.Errors.create(),
+    errors: new DSErrors(),
     username: null,
   });
 
   result = validate(undefined, undefined, model, 'username');
   assert.equal(processResult(result), true);
 
-  model.get('errors').add('username', 'Username is not unique');
+  model.get('errors').set('username', 'Username is not unique');
 
   result = validate(undefined, undefined, model, 'username');
   assert.equal(processResult(result), 'Username is not unique');
@@ -38,15 +50,15 @@ test('gets last message', function (assert) {
   assert.expect(2);
 
   model = EmberObject.create({
-    errors: DS.Errors.create(),
+    errors: new DSErrors(),
     username: null,
   });
 
   result = validate(undefined, undefined, model, 'username');
   assert.equal(processResult(result), true);
 
-  model.get('errors').add('username', 'Username is not unique');
-  model.get('errors').add('username', 'Username is too long');
+  model.get('errors').set('username', 'Username is not unique');
+  model.get('errors').set('username', 'Username is too long');
 
   result = validate(undefined, undefined, model, 'username');
   assert.equal(processResult(result), 'Username is too long');